Add KeychainStore and CompositeStore
We now first try the keychain then fall back to the JsonStore.
KeychainStore prefers the better-maintained keytar library, but
currently falls back on node-keychain if that fails for macOS
(as it does for me). Luckily keytar seems to only have problems
reading, so we're not putting passwords as cli arguments, just
reading them from stdout/stderr.