When loading txns for unbudgeted category, keep ONLY that category
Unbudgeted spending might have some transactions in a parent category
and some in a more specific child category; the total amount for these
were handled correctly, but we would include transactions for a child
category with the parent, leading to a confusing display. Now, we
try to show exclusively those transactions which contribute to the
total spending amount displayed (IE: those *exactly* in the given
category).

Note however that if you specifically budget a parent category, we
still include all child category transactions for budgeted spending,
since that's pretty obviously what you intended. For those with budgets
on both the parent category *and* one of its children... I'm not sure.
Maybe someone can contribute a PR if that's actually a thing.